[Resources] Relocate 6 non-user resources out of user_resources.py
Phase 1 of user_resources.py consolidation. Moves 6 resources from the
2,094-line user_resources.py (now 992 lines) into their natural sibling
modules, with zero behavior change to URIs, tags, descriptions, or
response shapes. All 13 resource URIs verified reachable post-refactor.

Relocations:
- tools://list/all, tools://detailed/list → tool_output_resources.py
- google://services/scopes/{service}      → service_list_resources.py
- template://user_email                   → template_resources.py
- workspace://content/search/{query}      → new workspace_resources.py
- gmail://allow-list                      → new gmail_resources.py

Also fixes a latent bug in workspace://content/search: the handler was
reading search_results.get("files") but search_drive_files returns its
list under the "results" key — the resource had always returned 0
total_results regardless of actual Drive matches until now.

Companion Auth changes bundled in:
- auth/middleware.py: track which extraction path resolved the user's
  identity (jwt/github_oauth/google_provider/session/oauth_file/tool_args)
  and persist via SessionKey.IDENTITY_SOURCE, so user://current/identity
  can surface the resolution source for dual-auth diagnostics.
- middleware/session_tool_filtering_middleware.py: redact email in the
  "Restored session" log line, extending the redaction pass from 95efe77.
